requirements.txt
asgiref==3.8.1
colorama==0.4.6
coverage==7.6.3
Django==5.1.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
iniconfig==2.0.0
Markdown==3.7
packaging==24.1
pillow==11.0.0
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.3
pytest-cov==5.0.0
pytest-django==4.9.0
sqlparse==0.5.1
tzdata==2024.2
